Rename package to main
Seems like package names are intended for libraries only. Go programs
that are intended to be executable all use "main" as the package name,
and Go refuses to build an executable otherwise.

If we wanted to make this work as a library as well as executable, we
could take an approach like golint and make a subfolder with the
executable that imports the library:
https://github.com/golang/lint/blob/master/golint/golint.go#L19